isA_common/requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
pydantic>=2.0.0
tenacity>=8.0.0
python-consul2>=0.1.5
